pyproject.toml
[project]
name = "demo-bank-customer-support"
version = "0.1.0"
description = "Demo Bank Customer Support"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.115.4",
    "python-dotenv>=1.0.1",
    "uvicorn>=0.32.0",
    "multinear>=0.1.1",
    "langchain-openai>=0.2.10",
    "langchain-community>=0.3.4",
    "langchain>=0.3.8",
    "faiss-cpu>=1.9.0.post1",
]
